+++ src/locales/fi/footerMessages.js
@@ -2,6 +2,7 @@
 // Licensed under the MIT License
 
 export default {
+  footerLogoImageText: 'Antin nimikirjoitus',
   footerSocialMediaTitle: 'Sosiaalisessa mediassa',
   footerFacebookImageText: 'Facebookin logo',
   footerGithubImageText: 'Githubin logo',